@@ -1,7 +1,7 @@
 package com.example.x.xalbum;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.widget.GridView;
 
 import com.example.x.xalbum.Adapter.AlbumAdapter;
