@@ -5,12 +5,12 @@ import android.content.Intent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.view.Window;
 import android.widget.BaseAdapter;
 import android.widget.ImageView;
 
 import com.bumptech.glide.Glide;
 import com.example.x.xalbum.Base.ImageData;
-import com.example.x.xalbum.ImageActivity;
 import com.example.x.xalbum.ImagePagerActivity;
 import com.example.x.xalbum.R;
 
